code/airflow_dock/requirements.txt
openai
geopandas 
shapely
beautifulsoup4
pandas
requests 
pathlib 
uuid6
python-dotenv 